# perf: use faster kmean find partition routing for pq assignment