--- a/rust/lance-index/Cargo.toml
+++ b/rust/lance-index/Cargo.toml
@@ -77,6 +77,10 @@ name = "pq_dist_table"
 harness = false
 
 [[bench]]
+name = "pq_assignment"
+harness = false
+
+[[bench]]
 name = "hnsw"
 harness = false