--- a/rust/lance-index/src/vector/pq.rs
+++ b/rust/lance-index/src/vector/pq.rs
@@ -14,7 +14,7 @@ use lance_arrow::*;
 use lance_core::{Error, Result};
 use lance_linalg::distance::{dot_distance_batch, l2_distance_batch, DistanceType, Dot, L2};
 use lance_linalg::kernels::argmin_value_float;
-use lance_linalg::kmeans::kmeans_find_partitions;
+use lance_linalg::kmeans::{compute_partitions, kmeans_find_partitions};
 use lance_linalg::{distance::MetricType, MatrixView};
 use rayon::prelude::*;
 use snafu::{location, Location};
@@ -360,12 +360,13 @@ where
                     location: location!(),
                 })?;
 
+        let sub_dim = dim / num_sub_vectors;
         let values = flatten_data
             .as_slice()
             .par_chunks(dim)
             .map(|vector| {
                 vector
-                    .chunks_exact(dim / num_sub_vectors)
+                    .chunks_exact(sub_dim)
                     .enumerate()
                     .flat_map(|(sub_idx, sub_vector)| {
                         let centroids = get_sub_vector_centroids(