--- a/rust/lance-index/src/vector/pq.rs
+++ b/rust/lance-index/src/vector/pq.rs
@@ -14,7 +14,7 @@ use lance_arrow::*;
 use lance_core::{Error, Result};
 use lance_linalg::distance::{dot_distance_batch, l2_distance_batch, DistanceType, Dot, L2};
 use lance_linalg::kernels::argmin_value_float;
-use lance_linalg::kmeans::compute_partitions;
+use lance_linalg::kmeans::compute_partition;
 use lance_linalg::{distance::MetricType, MatrixView};
 use rayon::prelude::*;
 use snafu::{location, Location};
@@ -368,7 +368,7 @@ where
                 vector
                     .chunks_exact(sub_dim)
                     .enumerate()
-                    .flat_map(|(sub_idx, sub_vector)| {
+                    .map(|(sub_idx, sub_vector)| {
                         let centroids = get_sub_vector_centroids(
                             codebook.as_slice(),
                             dim,
@@ -376,9 +376,7 @@ where
                             num_sub_vectors,
                             sub_idx,
                         );
-                        let parts =
-                            compute_partitions(centroids, sub_vector, sub_dim, distance_type);
-                        parts.iter().map(|v| v.unwrap() as u8).collect::<Vec<_>>()
+                        compute_partition(centroids, sub_vector, distance_type).map(|v| v as u8)
                     })
                     .collect::<Vec<_>>()
             })