--- a/rust/lance-index/src/vector/pq/utils.rs
+++ b/rust/lance-index/src/vector/pq/utils.rs
@@ -54,6 +54,7 @@ pub fn num_centroids(num_bits: impl Into<u32>) -> usize {
     2_usize.pow(num_bits.into())
 }
 
+#[inline]
 pub fn get_sub_vector_centroids<T>(
     codebook: &[T],
     dimension: usize,