* Added benchmark for pq assignment to use fast routing

```
Gnuplot not found, using plotters backend
l2,32768                time:   [321.08 ms 326.42 ms 331.92 ms]
                        change: [-51.461% -49.821% -48.361%] (p = 0.00 < 0.10)
                        Performance has improved.

Benchmarking dot,32768: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.4s.
dot,32768               time:   [514.21 ms 517.35 ms 520.75 ms]
                        change: [-40.887% -37.701% -34.616%] (p = 0.00 < 0.10)
                        Performance has improved.
```