@eddyxu 2024-06-23
```Gnuplot not found, using plotters backend
l2,32768                time:   [254.57 ms 258.13 ms 262.47 ms]
                        change: [-58.154% -57.389% -56.589%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

dot,32768               time:   [251.29 ms 258.31 ms 266.30 ms]
                        change: [-66.756% -65.430% -64.142%] (p = 0.00 < 0.10)
                        Performance has improved.
```